Move code from pipelines.js -> pipelineRepositories.js, etc.
